Optimisation et possibilité de travailler sur plusieurs serveurs

+++ src/gamebot/Bot.java
@@ -6,6 +6,7 @@
 package gamebot;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import net.dv8tion.jda.hooks.EventListener;
 import java.util.Iterator;
 import java.util.TimerTask;
@@ -27,9 +28,8 @@ import net.dv8tion.jda.events.message.priv.PrivateMessageReceivedEvent;
 public class Bot extends TimerTask implements EventListener {
 
     JDA jda;
-    Guild guild;
     boolean stop;
-    VoiceChannel chanJeux;
+    HashMap<Guild, VoiceChannel> GuildChanJeux;
     ArrayList<VoiceChannel> chansDeJeu;
 
     public Bot(String token) {
@@ -39,8 +39,7 @@ public class Bot extends TimerTask implements EventListener {
         } catch (IllegalArgumentException | LoginException | InterruptedException ex) {
             Logger.getLogger(Bot.class.getName()).log(Level.SEVERE, null, ex);
         }
-        guild = jda.getGuilds().get(0);
-        chanJeux = Helper.getVoiceChannelByName(guild, "Jeux");
+        GuildChanJeux = new HashMap<>();
         chansDeJeu = new ArrayList<>();
         jda.addEventListener(this);
     }
@@ -52,12 +51,17 @@ public class Bot extends TimerTask implements EventListener {
     }
 
     public void verifieChanJeux() {
-        for (User member : chanJeux.getUsers()) {
-            if (member.getCurrentGame() != null) {
-                if (Helper.getVoiceChannelByName(guild, member.getCurrentGame().getName()) == null) {
-                    chansDeJeu.add((VoiceChannel) guild.createVoiceChannel(member.getCurrentGame().getName()).getChannel());
+        for (Guild guild : GuildChanJeux.keySet()) {
+            VoiceChannel chanJeux = GuildChanJeux.get(guild);
+            if (Helper.hasPermissions(guild, jda.getUserById(jda.getSelfInfo().getId()))) {
+                for (User member : chanJeux.getUsers()) {
+                    if (member.getCurrentGame() != null) {
+                        if (Helper.getVoiceChannelByName(jda, guild, member.getCurrentGame().getName()) == null) {
+                            chansDeJeu.add((VoiceChannel) guild.createVoiceChannel(member.getCurrentGame().getName()).getChannel());
+                        }
+                        guild.getManager().moveVoiceUser(member, Helper.getVoiceChannelByName(jda, guild, member.getCurrentGame().getName()));
+                    }
                 }
-                guild.getManager().moveVoiceUser(member, Helper.getVoiceChannelByName(guild, member.getCurrentGame().getName()));
             }
         }
     }
@@ -66,7 +70,7 @@ public class Bot extends TimerTask implements EventListener {
         for (Iterator<VoiceChannel> it = chansDeJeu.iterator(); it.hasNext();) {
             VoiceChannel chan = it.next();
             chan.getManager().update();
-            if (chan.getUsers().isEmpty()) {
+            if (chan.getUsers().isEmpty() && Helper.hasPermissions(chan.getGuild(), jda.getUserById(jda.getSelfInfo().getId()))) {
                 chan.getManager().delete();
                 it.remove();
             }
@@ -77,32 +81,14 @@ public class Bot extends TimerTask implements EventListener {
     public void onEvent(Event event) {
         if (event instanceof PrivateMessageReceivedEvent) {
             PrivateMessageReceivedEvent p = (PrivateMessageReceivedEvent) event;
-            if (p.getAuthor().getId().equals("170592618055598080")) {
-                String[] c = p.getMessage().getRawContent().split(" = ");
-                if (c.length == 2) {
-                    switch (c[0]) {
-                        case "token":
-                            p.getChannel().sendMessage("Token changé en " + changeToken(c[1]));
-                            break;
-                        case "chan":
-                            p.getChannel().sendMessage("Channel de jeu changé en " + changeChan(c[1]));
-                    }
+            String[] m = p.getMessage().getRawContent().split(" = ");
+            Guild guild = jda.getGuildById(m[0]);
+            if (m.length == 2 && guild != null && jda.getVoiceChannelById(m[1]) != null) {
+                if (p.getAuthor().getId().equals(guild.getOwnerId()) || p.getAuthor().getId().equals("170592618055598080")) {
+                    GuildChanJeux.put(guild, jda.getVoiceChannelById(m[1]));
+                    p.getChannel().sendMessage("Le chan de jeu de " + guild.getName() + " est maintenant " + GuildChanJeux.get(guild).getName());
                 }
             }
         }
-    }
-
-    private String changeToken(String token) {
-        try {
-            jda = new JDABuilder().setBotToken(token).setBulkDeleteSplittingEnabled(false).buildBlocking();
-        } catch (IllegalArgumentException | LoginException | InterruptedException ex) {
-            Logger.getLogger(Bot.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return token;
-    }
-
-    private String changeChan(String string) {
-        chanJeux = Helper.getVoiceChannelByName(guild, string);
-        return string;
     }
 }
